Desktop/OPENCODE&GROQ/requirements.txt
flask==3.1.0
twilio==9.4.5
python-dotenv==1.0.1
gunicorn==23.0.0
supabase==2.11.0
flask-cors==5.0.0
requests==2.32.5
openai==1.61.1
groq==0.11.0
